# internal/httpapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "dflockd HTTP API",
    "version": "2.1.0",
    "summary": "REST surface for dflockd's distributed FIFO lock and semaphore primitives.",
...
  },
  "servers": [
    {
      "url": "http://{host}:{port}",
      "description": "Configured HTTP API listener",
      "variables": {
        "host": {"default": "127.0.0.1"},
        "port": {"default": "6389"}
      }
    }
  ],
  "security": [{"bearerAuth": []}],
...
  "paths": {
...
    "/v1/sessions": {
      "post": {
        "tags": ["sessions"],
        "summary": "Create a session",
        "operationId": "createSession",
        "description": "Mints a new session with a server-generated 32-character lowercase hex ID. The response advertises the configured idle timeout so clients know how often to ping.",
        "responses": {
          "200": {"description": "Created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateSessionResponse"}, "example": {"session_id": "0fe1f0c5b6758b10fe1ce7df57c012ef", "idle_timeout_s": 20}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "503": {"description": "Cap reached or draining", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorBody"}, "examples": {"max_sessions": {"value": {"error": "max_sessions"}}, "max_sessions_per_ip": {"value": {"error": "max_sessions_per_ip"}}, "draining": {"value": {"error": "draining"}}}}}}
        }
      }
    },
    "/v1/sessions/{id}": {
      "delete": {
        "tags": ["sessions"],
        "summary": "Delete a session",
        "operationId": "deleteSession",
        "description": "Synchronously closes the session. Held locks and semaphore slots are released. Idempotent; deleting an unknown id returns 410.",
        "parameters": [{"$ref": "#/components/parameters/SessionIdPath"}],
        "responses": {
          "204": {"description": "Closed"},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "410": {"$ref": "#/components/responses/SessionGone"},
          "503": {"$ref": "#/components/responses/CapacityExhausted"}
        }
      }
    },
...
    "/v1/stats": {
      "get": {
        "tags": ["introspection"],
        "summary": "LockManager snapshot",
        "operationId": "getStats",
        "description": "Returns active connections, held locks/semaphores with owner/lease info, and idle resources awaiting GC.",
        "responses": {
          "200": {"description": "Snapshot", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Stats"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"}
        }
      }
    },
    "/v1/readindex": {
      "get": {
        "tags": ["cluster"],
        "summary": "Linearizable read barrier",
        "operationId": "readIndex",
        "description": "Proposes a no-op through Raft and returns 200 once it applies. After the call returns, every preceding committed write is reflected in state observable on this leader. On a follower returns 503 with X-Dflockd-Leader; on a single-node server returns 404.",
        "responses": {
          "200": {"description": "Barrier applied", "content": {"application/json": {"schema": {"type": "object", "properties": {"status": {"type": "string", "enum": ["ok"]}}}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "404": {"description": "Not in cluster mode", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorBody"}, "example": {"error": "cluster_disabled"}}}},
          "503": {"description": "Not leader (X-Dflockd-Leader header carries the leader's raft client address)", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorBody"}, "example": {"error": "not_leader"}}}}
        }
      }
    },
    "/v1/admin/voters": {
      "post": {
        "tags": ["admin"],
        "summary": "Add a voting cluster member",
        "operationId": "adminAddVoter",
        "description": "Proposes a single-server addition through Raft. Requires X-Dflockd-Admin (constant-time compared against --admin-token / DFLOCKD_ADMIN_TOKEN). 503 admin_disabled when no admin token is configured; 503 not_leader on a follower (X-Dflockd-Leader carries the leader address).",
        "parameters": [
          {"name": "X-Dflockd-Admin", "in": "header", "required": true, "schema": {"type": "string"}, "description": "Admin secret. Constant-time compared."}
        ],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"type": "object", "required": ["node_id", "raft_addr", "client_addr"], "properties": {"node_id": {"type": "string"}, "raft_addr": {"type": "string", "example": "10.0.0.4:7001"}, "client_addr": {"type": "string", "example": "10.0.0.4:6388"}}}}}
        },
        "responses": {
          "200": {"description": "Voter added", "content": {"application/json": {"schema": {"type": "object", "properties": {"status": {"type": "string"}, "node_id": {"type": "string"}}}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"description": "Wrong admin token", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorBody"}, "example": {"error": "admin_unauthorized"}}}},
          "503": {"description": "Either admin disabled (no token) or this node is a follower", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorBody"}}}}
        }
      }
    },
    "/v1/admin/voters/{id}": {
      "delete": {
        "tags": ["admin"],
        "summary": "Remove a voting cluster member",
        "operationId": "adminRemoveVoter",
        "description": "Proposes a single-server removal through Raft. Same auth model as POST /v1/admin/voters. A leader that removes itself steps down once the entry commits.",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "string"}, "description": "Stable node id."},
          {"name": "X-Dflockd-Admin", "in": "header", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Voter removed", "content": {"application/json": {"schema": {"type": "object", "properties": {"status": {"type": "string"}, "node_id": {"type": "string"}}}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"description": "Wrong admin token", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorBody"}}}},
          "503": {"description": "Either admin disabled (no token) or this node is a follower", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorBody"}}}}
        }
      }
    },
    "/v1/locks/{key}": {
      "post": {
        "tags": ["locks"],
        "summary": "Acquire a lock",
        "operationId": "acquireLock",
        "description": "Single-phase acquire. Blocks server-side up to `acquire_timeout_s` (0 = non-blocking try-lock).",
        "parameters": [{"$ref": "#/components/parameters/KeyPath"}, {"$ref": "#/components/parameters/SessionHeader"}],
        "requestBody": {"$ref": "#/components/requestBodies/AcquireLock"},
        "responses": {
          "200": {"$ref": "#/components/responses/AcquireOutcome"},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "408": {"$ref": "#/components/responses/ClientCanceled"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "410": {"$ref": "#/components/responses/SessionGone"},
          "503": {"$ref": "#/components/responses/CapacityExhausted"}
        }
      }
    },
    "/v1/locks/{key}/release": {
      "post": {
        "tags": ["locks"],
        "summary": "Release a held lock",
        "operationId": "releaseLock",
        "parameters": [{"$ref": "#/components/parameters/KeyPath"}, {"$ref": "#/components/parameters/SessionHeader"}],
        "requestBody": {"$ref": "#/components/requestBodies/Release"},
        "responses": {
          "204": {"description": "Released"},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "404": {"description": "Token not held", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorBody"}, "example": {"error": "not_held"}}}},
          "410": {"$ref": "#/components/responses/SessionGone"},
          "503": {"$ref": "#/components/responses/CapacityExhausted"}
        }
      }
    },
    "/v1/locks/{key}/renew": {
      "post": {
        "tags": ["locks"],
        "summary": "Renew a held lock's lease",
        "operationId": "renewLock",
        "parameters": [{"$ref": "#/components/parameters/KeyPath"}, {"$ref": "#/components/parameters/SessionHeader"}],
        "requestBody": {"$ref": "#/components/requestBodies/Renew"},
        "responses": {
          "200": {"description": "Renewed", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RenewResponse"}, "example": {"remaining_s": 60}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "404": {"description": "Token not held", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorBody"}, "example": {"error": "not_held"}}}},
          "410": {"$ref": "#/components/responses/SessionGone"},
          "503": {"$ref": "#/components/responses/CapacityExhausted"}
        }
      }
    },
    "/v1/locks/{key}/enqueue": {
      "post": {
        "tags": ["locks"],
        "summary": "Two-phase enqueue (phase 1)",
        "operationId": "enqueueLock",
        "description": "Either grants the lock immediately (`acquired`) or registers a waiter (`queued`). Pair with `/wait` to block until grant.",
        "parameters": [{"$ref": "#/components/parameters/KeyPath"}, {"$ref": "#/components/parameters/SessionHeader"}],
        "requestBody": {"$ref": "#/components/requestBodies/Enqueue"},
        "responses": {
          "200": {"$ref": "#/components/responses/EnqueueOutcome"},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "410": {"$ref": "#/components/responses/SessionGone"},
          "503": {"$ref": "#/components/responses/CapacityExhausted"}
        }
      }
    },
    "/v1/locks/{key}/wait": {
      "post": {
        "tags": ["locks"],
        "summary": "Two-phase wait (phase 2)",
        "operationId": "waitLock",
        "description": "Blocks server-side up to `timeout_s` waiting for the grant from a prior `enqueue` on the same session.",
        "parameters": [{"$ref": "#/components/parameters/KeyPath"}, {"$ref": "#/components/parameters/SessionHeader"}],
        "requestBody": {"$ref": "#/components/requestBodies/Wait"},
        "responses": {
          "200": {"$ref": "#/components/responses/AcquireOutcome"},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "408": {"$ref": "#/components/responses/ClientCanceled"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "410": {"$ref": "#/components/responses/SessionGone"},
          "503": {"$ref": "#/components/responses/CapacityExhausted"}
        }
      }
    },
    "/v1/semaphores/{key}": {
      "post": {
        "tags": ["semaphores"],
        "summary": "Acquire a semaphore slot",
        "operationId": "acquireSem",
        "description": "Single-phase acquire. The first acquirer for a key sets the `limit`; subsequent callers must match it or get `409 limit_mismatch`.",
        "parameters": [{"$ref": "#/components/parameters/KeyPath"}, {"$ref": "#/components/parameters/SessionHeader"}],
        "requestBody": {"$ref": "#/components/requestBodies/AcquireSem"},
        "responses": {
          "200": {"$ref": "#/components/responses/AcquireOutcome"},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "408": {"$ref": "#/components/responses/ClientCanceled"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "410": {"$ref": "#/components/responses/SessionGone"},
          "503": {"$ref": "#/components/responses/CapacityExhausted"}
        }
      }
    },
    "/v1/semaphores/{key}/release": {
      "post": {
        "tags": ["semaphores"],
        "summary": "Release a semaphore slot",
        "operationId": "releaseSem",
        "parameters": [{"$ref": "#/components/parameters/KeyPath"}, {"$ref": "#/components/parameters/SessionHeader"}],
        "requestBody": {"$ref": "#/components/requestBodies/Release"},
        "responses": {
          "204": {"description": "Released"},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "404": {"description": "Token not held", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorBody"}, "example": {"error": "not_held"}}}},
          "410": {"$ref": "#/components/responses/SessionGone"},
          "503": {"$ref": "#/components/responses/CapacityExhausted"}
        }
      }
    },
    "/v1/semaphores/{key}/renew": {
      "post": {
        "tags": ["semaphores"],
        "summary": "Renew a semaphore slot's lease",
        "operationId": "renewSem",
        "parameters": [{"$ref": "#/components/parameters/KeyPath"}, {"$ref": "#/components/parameters/SessionHeader"}],
        "requestBody": {"$ref": "#/components/requestBodies/Renew"},
        "responses": {
          "200": {"description": "Renewed", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RenewResponse"}, "example": {"remaining_s": 60}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "404": {"description": "Token not held", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorBody"}, "example": {"error": "not_held"}}}},
          "410": {"$ref": "#/components/responses/SessionGone"},
          "503": {"$ref": "#/components/responses/CapacityExhausted"}
        }
      }
    },
    "/v1/semaphores/{key}/enqueue": {
      "post": {
        "tags": ["semaphores"],
        "summary": "Two-phase semaphore enqueue (phase 1)",
        "operationId": "enqueueSem",
        "parameters": [{"$ref": "#/components/parameters/KeyPath"}, {"$ref": "#/components/parameters/SessionHeader"}],
        "requestBody": {"$ref": "#/components/requestBodies/EnqueueSem"},
        "responses": {
          "200": {"$ref": "#/components/responses/EnqueueOutcome"},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "410": {"$ref": "#/components/responses/SessionGone"},
          "503": {"$ref": "#/components/responses/CapacityExhausted"}
        }
      }
    },
    "/v1/semaphores/{key}/wait": {
      "post": {
        "tags": ["semaphores"],
        "summary": "Two-phase semaphore wait (phase 2)",
        "operationId": "waitSem",
        "parameters": [{"$ref": "#/components/parameters/KeyPath"}, {"$ref": "#/components/parameters/SessionHeader"}],
        "requestBody": {"$ref": "#/components/requestBodies/Wait"},
        "responses": {
          "200": {"$ref": "#/components/responses/AcquireOutcome"},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "408": {"$ref": "#/components/responses/ClientCanceled"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "410": {"$ref": "#/components/responses/SessionGone"},
          "503": {"$ref": "#/components/responses/CapacityExhausted"}
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer", "description": "Shared-secret token configured via `--auth-token` / `--auth-token-file` / `DFLOCKD_AUTH_TOKEN[_FILE]`."}
    },
    "parameters": {
      "KeyPath": {
        "name": "key",
        "in": "path",
        "required": true,
        "description": "Lock or semaphore key. Must not be empty or contain whitespace; max 256 bytes.",
        "schema": {"type": "string", "minLength": 1, "maxLength": 256, "pattern": "^\\S+$"}
      },
      "SessionIdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Session id (32 lowercase hex chars).",
        "schema": {"type": "string", "pattern": "^[0-9a-f]{32}$"}
      },
      "SessionHeader": {
        "name": "X-Dflockd-Session",
        "in": "header",
        "required": true,
        "description": "Session id from `POST /v1/sessions`. Required on every lock-modifying request.",
        "schema": {"type": "string", "pattern": "^[0-9a-f]{32}$"}
      }
    },
    "requestBodies": {
      "AcquireLock": {
        "required": true,
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AcquireRequest"}, "example": {"acquire_timeout_s": 10, "lease_ttl_s": 60}}}
      },
      "AcquireSem": {
        "required": true,
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SemAcquireRequest"}, "example": {"acquire_timeout_s": 10, "limit": 5, "lease_ttl_s": 60}}}
      },
      "Release": {
        "required": true,
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ReleaseRequest"}, "example": {"token": "0001a3f217b3c4d87f3c1f2b3e9a8d6e"}}}
      },
      "Renew": {
        "required": true,
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RenewRequest"}, "example": {"token": "0001a3f217b3c4d87f3c1f2b3e9a8d6e", "lease_ttl_s": 60}}}
      },
      "Enqueue": {
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/EnqueueRequest"}, "example": {"lease_ttl_s": 60}}}
      },
      "EnqueueSem": {
        "required": true,
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SemEnqueueRequest"}, "example": {"limit": 5, "lease_ttl_s": 60}}}
      },
      "Wait": {
        "required": true,
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/WaitRequest"}, "example": {"timeout_s": 30}}}
      }
    },
    "responses": {
      "BadRequest": {"description": "Malformed request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorBody"}, "example": {"error": "bad_request", "detail": "X-Dflockd-Session header required"}}}},
      "Unauthorized": {"description": "Missing or invalid bearer token", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorBody"}, "example": {"error": "unauthorized"}}}},
      "ClientCanceled": {"description": "Client disconnected during a long-poll", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorBody"}, "example": {"error": "client_canceled"}}}},
      "Conflict": {
        "description": "Operation conflicts with existing state",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorBody"}, "examples": {
          "limit_mismatch": {"value": {"error": "limit_mismatch"}},
          "already_enqueued": {"value": {"error": "already_enqueued"}},
          "not_enqueued": {"value": {"error": "not_enqueued"}},
          "lease_expired": {"value": {"error": "lease_expired"}}
        }}}
      },
      "SessionGone": {"description": "Session unknown, expired, or DELETEd", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorBody"}, "example": {"error": "session_gone"}}}},
      "CapacityExhausted": {
        "description": "Server-side capacity or fence persistence unavailable",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorBody"}, "examples": {
          "max_locks": {"value": {"error": "max_locks"}},
          "max_waiters": {"value": {"error": "max_waiters"}},
          "fence_persistence": {"value": {"error": "fence_persistence"}},
          "draining": {"value": {"error": "draining"}}
        }}}
      },
      "AcquireOutcome": {
        "description": "Acquire/wait result",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OpResponse"}, "examples": {
          "granted": {"value": {"status": "ok", "token": "0001a3f217b3c4d87f3c1f2b3e9a8d6e", "lease_ttl_s": 60}},
          "timeout": {"value": {"status": "timeout"}}
        }}}
      },
      "EnqueueOutcome": {
        "description": "Enqueue result",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OpResponse"}, "examples": {
          "acquired": {"value": {"status": "acquired", "token": "0001a3f217b3c4d87f3c1f2b3e9a8d6e", "lease_ttl_s": 60}},
          "queued": {"value": {"status": "queued"}}
        }}}
      }
    },
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {"type": "string", "description": "Stable machine-readable code."},
          "detail": {"type": "string", "description": "Optional human-readable note."}
        }
      },
      "CreateSessionResponse": {
        "type": "object",
        "required": ["session_id", "idle_timeout_s"],
        "properties": {
          "session_id": {"type": "string", "pattern": "^[0-9a-f]{32}$"},
          "idle_timeout_s": {"type": "integer", "minimum": 0, "description": "Advisory; ping at half this interval to be safe."}
        }
      },
...
      "RenewResponse": {
        "type": "object",
        "required": ["remaining_s"],
        "properties": {"remaining_s": {"type": "integer", "minimum": 0}}
      },
      "Stats": {
        "type": "object",
        "required": ["connections", "locks", "semaphores", "idle_locks", "idle_semaphores"],
        "properties": {
          "connections": {"type": "integer", "description": "Active TCP connections plus active HTTP sessions."},
          "locks": {"type": "array", "items": {"$ref": "#/components/schemas/LockInfo"}},
          "semaphores": {"type": "array", "items": {"$ref": "#/components/schemas/SemInfo"}},
          "idle_locks": {"type": "array", "items": {"$ref": "#/components/schemas/IdleInfo"}},
          "idle_semaphores": {"type": "array", "items": {"$ref": "#/components/schemas/IdleInfo"}}
        }
      },
      "LockInfo": {
        "type": "object",
        "required": ["key"],
        "properties": {
          "key": {"type": "string"},
          "owner_conn_id": {"type": "integer", "description": "0 if there's no current holder."},
          "lease_expires_in_s": {"type": "number"},
          "waiters": {"type": "integer", "minimum": 0}
        }
      },
...
    }
  }
}